+++ RemoteSwitch/RemoteTransmitter.cpp
@@ -12,7 +12,7 @@
 * RemoteTransmitter
 ************/
 
-RemoteTransmitter::RemoteTransmitter(unsigned short pin, unsigned int periodusec, unsigned short repeats) {
+RemoteTransmitter::RemoteTransmitter(byte pin, unsigned int periodusec, byte repeats) {
 	_pin=pin;
 	_periodusec=periodusec;
 	_repeats=repeats;
@@ -20,11 +20,11 @@ RemoteTransmitter::RemoteTransmitter(unsigned short pin, unsigned int periodusec
 	pinMode(_pin, OUTPUT);
 }
 
-unsigned long RemoteTransmitter::encodeTelegram(unsigned short trits[]) {
+unsigned long RemoteTransmitter::encodeTelegram(byte trits[]) {
 	unsigned long data = 0;
 
 	// Encode data
-	for (unsigned short i=0;i<12;i++) {
+	for (byte i=0;i<12;i++) {
 		data*=3;
 		data+=trits[i];
 	}
@@ -38,7 +38,7 @@ unsigned long RemoteTransmitter::encodeTelegram(unsigned short trits[]) {
 	return data;
 }
 
-void RemoteTransmitter::sendTelegram(unsigned short trits[]) {
+void RemoteTransmitter::sendTelegram(byte trits[]) {
 	sendTelegram(encodeTelegram(trits),_pin);
 }
 
@@ -49,21 +49,21 @@ void RemoteTransmitter::sendTelegram(unsigned short trits[]) {
 * r = repeats as 2log. Thus, if r = 3, then signal is sent 2^3=8 times
 * d = data
 */
-void RemoteTransmitter::sendTelegram(unsigned long data, unsigned short pin) {
+void RemoteTransmitter::sendTelegram(unsigned long data, byte pin) {
 	unsigned int periodusec = (unsigned long)data >> 23;
-	unsigned short repeats = ((unsigned long)data >> 20) & B111;
+	byte repeats = ((unsigned long)data >> 20) & B111;
 
 	sendCode(pin, data, periodusec, repeats);
 }
 
-void RemoteTransmitter::sendCode(unsigned short pin, unsigned long code, unsigned int periodusec, unsigned short repeats) {
+void RemoteTransmitter::sendCode(byte pin, unsigned long code, unsigned int periodusec, byte repeats) {
 	code &= 0xfffff; // Truncate to 20 bit ;
 	// Convert the base3-code to base4, to avoid lengthy calculations when transmitting.. Messes op timings.
 	// Also note this swaps endianess in the process. The MSB must be transmitted first, but is converted to
 	// LSB here. This is easier when actually transmitting later on.
 	unsigned long dataBase4 = 0;
 
-	for (unsigned short i=0; i<12; i++) {
+	for (byte i=0; i<12; i++) {
 		dataBase4<<=2;
 		dataBase4|=(code%3);
 		code/=3;
@@ -71,12 +71,12 @@ void RemoteTransmitter::sendCode(unsigned short pin, unsigned long code, unsigne
 
 	repeats = 1 << (repeats & B111); // repeats := 2^repeats;
 
-	for (unsigned short int j=0;j<repeats;j++) {
+	for (byte j=0;j<repeats;j++) {
 		// Sent one telegram
 
 		// Recycle code as working var to save memory
 		code=dataBase4;
-		for (unsigned short i=0; i<12; i++) {
+		for (byte i=0; i<12; i++) {
 			switch (code & B11) {
 				case 0:
 					digitalWrite(pin, HIGH);
@@ -113,7 +113,7 @@ void RemoteTransmitter::sendCode(unsigned short pin, unsigned long code, unsigne
 			code>>=2;
 		}
 
-		// Send termination/synchronisation-signal. Total length: 32 periods
+		// Send termination/synchronization-signal. Total length: 32 periods
 		digitalWrite(pin, HIGH);
 		delayMicroseconds(periodusec);
 		digitalWrite(pin, LOW);
@@ -122,7 +122,7 @@ void RemoteTransmitter::sendCode(unsigned short pin, unsigned long code, unsigne
 }
 
 boolean RemoteTransmitter::isSameCode(unsigned long encodedTelegram, unsigned long receivedData) {
-	return (receivedData==(encodedTelegram & 0xFFFFF)); // ompare the 20 LSB's
+	return (receivedData==(encodedTelegram & 0xFFFFF)); // compare the 20 LSB's
 }
 
 
@@ -130,21 +130,21 @@ boolean RemoteTransmitter::isSameCode(unsigned long encodedTelegram, unsigned lo
 * ActionTransmitter
 ************/
 
-ActionTransmitter::ActionTransmitter(unsigned short pin, unsigned int periodusec, unsigned short repeats) : RemoteTransmitter(pin,periodusec,repeats) {
-	// Call contructor
+ActionTransmitter::ActionTransmitter(byte pin, unsigned int periodusec, byte repeats) : RemoteTransmitter(pin,periodusec,repeats) {
+	// Call constructor
 }
 
 
-void ActionTransmitter::sendSignal(unsigned short systemCode, char device, boolean on) {
+void ActionTransmitter::sendSignal(byte systemCode, char device, boolean on) {
 	sendTelegram(getTelegram(systemCode,device,on), _pin);
 }
 
-unsigned long ActionTransmitter::getTelegram(unsigned short systemCode, char device, boolean on) {
-	unsigned short trits[12];
+unsigned long ActionTransmitter::getTelegram(byte systemCode, char device, boolean on) {
+	byte trits[12];
 
 	device-=65;
 
-	for (unsigned short i=0; i<5; i++) {
+	for (byte i=0; i<5; i++) {
 		// Trits 0-4 contain address (2^5=32 addresses)
 		trits[i]=(systemCode & 1)?1:2;
 		systemCode>>=1;
@@ -164,21 +164,21 @@ unsigned long ActionTransmitter::getTelegram(unsigned short systemCode, char dev
 * BlokkerTransmitter
 ************/
 
-BlokkerTransmitter::BlokkerTransmitter(unsigned short pin, unsigned int periodusec, unsigned short repeats) : RemoteTransmitter(pin,periodusec,repeats) {
-	// Call contructor
+BlokkerTransmitter::BlokkerTransmitter(byte pin, unsigned int periodusec, byte repeats) : RemoteTransmitter(pin,periodusec,repeats) {
+	// Call constructor
 }
 
 
-void BlokkerTransmitter::sendSignal(unsigned short device, boolean on) {
+void BlokkerTransmitter::sendSignal(byte device, boolean on) {
 	sendTelegram(getTelegram(device,on), _pin);
 }
 
-unsigned long BlokkerTransmitter::getTelegram(unsigned short device, boolean on) {
-	unsigned short trits[12]={0};
+unsigned long BlokkerTransmitter::getTelegram(byte device, boolean on) {
+	byte trits[12]={0};
 
 	device--;
 
-	for (unsigned short i=1; i<4; i++) {
+	for (byte i=1; i<4; i++) {
 		// Trits 1-3 contain device
 		trits[i]=(device & 1)?0:1;
 		device>>=1;
@@ -194,21 +194,21 @@ unsigned long BlokkerTransmitter::getTelegram(unsigned short device, boolean on)
 * KaKuTransmitter
 ************/
 
-KaKuTransmitter::KaKuTransmitter(unsigned short pin, unsigned int periodusec, unsigned short repeats) : RemoteTransmitter(pin,periodusec,repeats) {
-	// Call contructor
+KaKuTransmitter::KaKuTransmitter(byte pin, unsigned int periodusec, byte repeats) : RemoteTransmitter(pin,periodusec,repeats) {
+	// Call constructor
 }
 
-void KaKuTransmitter::sendSignal(char address, unsigned short device, boolean on) {
+void KaKuTransmitter::sendSignal(char address, byte device, boolean on) {
 	sendTelegram(getTelegram(address, device, on), _pin);
 }
 
-unsigned long KaKuTransmitter::getTelegram(char address, unsigned short device, boolean on) {
-	unsigned short trits[12];
+unsigned long KaKuTransmitter::getTelegram(char address, byte device, boolean on) {
+	byte trits[12];
 
 	address-=65;
 	device-=1;
 
-	for (unsigned short i=0; i<4; i++) {
+	for (byte i=0; i<4; i++) {
 		// Trits 0-3 contain address (2^4 = 16 addresses)
 		trits[i]=(address & 1)?2:0;
 		address>>=1;
@@ -229,12 +229,12 @@ unsigned long KaKuTransmitter::getTelegram(char address, unsigned short device, 
 	return encodeTelegram(trits);
 }
 
-void KaKuTransmitter::sendSignal(char address, unsigned short group, unsigned short device, boolean on) {
+void KaKuTransmitter::sendSignal(char address, byte group, byte device, boolean on) {
 	sendTelegram(getTelegram(address, group, on), _pin);
 }
 
-unsigned long KaKuTransmitter::getTelegram(char address, unsigned short group, unsigned short device, boolean on) {
-	unsigned short trits[12], i;
+unsigned long KaKuTransmitter::getTelegram(char address, byte group, byte device, boolean on) {
+	byte trits[12], i;
 
 	address-=65;
 	group-=1;
@@ -275,20 +275,20 @@ unsigned long KaKuTransmitter::getTelegram(char address, unsigned short group, u
 * ElroTransmitter
 ************/
 
-ElroTransmitter::ElroTransmitter(unsigned short pin, unsigned int periodusec, unsigned short repeats) : RemoteTransmitter(pin, periodusec, repeats) {
-	//Call contructor
+ElroTransmitter::ElroTransmitter(byte pin, unsigned int periodusec, byte repeats) : RemoteTransmitter(pin, periodusec, repeats) {
+	//Call constructor
 }
 
-void ElroTransmitter::sendSignal(unsigned short systemCode, char device, boolean on) {
+void ElroTransmitter::sendSignal(byte systemCode, char device, boolean on) {
 	sendTelegram(getTelegram(systemCode, device, on), _pin);
 }
 
-unsigned long ElroTransmitter::getTelegram(unsigned short systemCode, char device, boolean on) {
-	unsigned short trits[12];
+unsigned long ElroTransmitter::getTelegram(byte systemCode, char device, boolean on) {
+	byte trits[12];
 
 	device-=65;
 
-	for (unsigned short i=0; i<5; i++) {
+	for (byte i=0; i<5; i++) {
 		//trits 0-4 contain address (2^5=32 addresses)
 		trits[i]=(systemCode & 1)?0:2;
 		systemCode>>=1;
